+++ extensions/codex/src/app-server/transcript-mirror.ts
@@ -49,8 +49,6 @@ type CodexThreadHistoryImportResult = {
   omittedMessages: number;
 };
 
-export type CodexThreadHistory = Pick<CodexThread, "id" | "createdAt" | "modelProvider" | "turns">;
-
 type BoundedCodexThreadHistoryProjection = CodexThreadHistoryImportResult & {
   responseItems: JsonValue[];
   transcriptMessages: AgentMessage[];
@@ -126,7 +124,7 @@ function projectCodexUserItemText(item: Record<string, unknown>): string | undef
 }
 
 function selectTurnsThroughBoundary(
-  thread: CodexThreadHistory,
+  thread: CodexThread,
   throughTurnId: string | null,
 ): NonNullable<CodexThread["turns"]> {
   if (throughTurnId === null) {
@@ -149,7 +147,7 @@ function selectTurnsThroughBoundary(
 }
 
 function projectCodexThreadHistory(params: {
-  thread: CodexThreadHistory;
+  thread: CodexThread;
   throughTurnId: string | null;
   importedAt: number;
   modelProvider?: string;
@@ -254,7 +252,7 @@ function selectBoundedCodexHistoryTail(
 
 /** Projects one terminal Codex history prefix into transcript and Responses API items. */
 export function projectBoundedCodexThreadHistory(params: {
-  thread: CodexThreadHistory;
+  thread: CodexThread;
   throughTurnId: string | null;
   importedAt: number;
   modelProvider?: string | null;
@@ -276,7 +274,7 @@ export function projectBoundedCodexThreadHistory(params: {
 
 /** Imports a bounded, user-visible Codex history tail into a new OpenClaw transcript. */
 export async function importCodexThreadHistoryToTranscript(params: {
-  thread: CodexThreadHistory;
+  thread: CodexThread;
   throughTurnId: string | null;
   storePath: string;
   sessionId: string;
